added display module
