+++ board.rb
@@ -1,0 +1,8 @@
+class Board
+  def initialize
+    @board = Array.new(12){Array.new(4)}
+  end
+
+
+
+end
