+++ breaker.rb
@@ -1,0 +1,7 @@
+class Breaker < Player
+  def guess
+    get_code_input
+    make guess
+
+  end
+end
