+++ feedback_board.rb
@@ -1,0 +1,3 @@
+class FeedBackBoard < Board
+
+end
