+++ maker.rb
@@ -1,0 +1,11 @@
+class Maker < Player
+  def set_code
+    get_code_input
+    set game code
+
+  end
+
+  def feedback
+
+  end
+end
