+++ mastermind.rb
@@ -33,53 +33,3 @@ class Mastermind
   end
 
 end
-
-
-class Player
-
-    def initialize(board)
-      @board = board
-    end
-
-    def get_code_input
-
-    end
-
-end
-
-class Maker < Player
-  def set_code
-    get_code_input
-    set game code
-
-  end
-
-  def feedback
-
-  end
-end
-
-class Breaker < Player
-  def guess
-    get_code_input
-    make guess
-
-  end
-end
-
-class Board
-  def initialize
-    @arr =[]
-    12.times do
-      arr2=[]
-      4.times {arr2 << "-"}
-      @arr << arr
-    end
-  end
-
-
-end
-
-class FeedBackBoard < Board
-
-end
