+++ player.rb
@@ -1,0 +1,11 @@
+class Player
+
+    def initialize(board)
+      @board = board
+    end
+
+    def get_code_input
+
+    end
+
+end
